+++ isik/django/drf/viewsets/__init__.py
@@ -2,7 +2,7 @@ from isik.django.drf.viewsets.action_serializer_class import ActionSerializerCla
 from isik.django.drf.viewsets.base import BaseModelViewSet
 from isik.django.drf.viewsets.filterset import FilterSetMixin
 from isik.django.drf.viewsets.history import HistoryMixin, context_filter
-from isik.django.drf.viewsets.ordering import ReverseOrderingMixin
+from isik.django.drf.viewsets.ordering import DeclaredOrderingFilter, DeclaredOrderingMixin, ReverseOrderingMixin
 from isik.django.drf.viewsets.protected_destroy import ProtectedDestroyMixin
 from isik.django.drf.viewsets.registry import ViewSetRegistryMixin
 from isik.django.drf.viewsets.schema_generation import none_during_schema_generation
@@ -11,6 +11,8 @@ from isik.django.drf.viewsets.schema_generation import none_during_schema_genera
 __all__ = [
     "ActionSerializerClassMixin",
     "BaseModelViewSet",
+    "DeclaredOrderingFilter",
+    "DeclaredOrderingMixin",
     "FilterSetMixin",
     "HistoryMixin",
     "ProtectedDestroyMixin",